requirements.txt
nltk==3.3
numpy==1.15.4
torch==0.4.1.post2
spacy==2.0.16
attrs==18.1
allennlp==0.7.1
